requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
boto3==1.24.13
botocore==1.27.13
certifi==2021.10.8
cffi==1.15.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==3.0.1
django-storages==1.12.3
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
gunicorn==20.1.0
ipykernel==6.9.2
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
jmespath==1.0.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pipenv==2022.1.8
platformdirs==2.5.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
s3transfer==0.6.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
virtualenv==20.13.3
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.0.0
widgetsnbextension==3.6.0
